Revert "compile"
This reverts commit 0e9bfe8e88e2f49d46f0ebb94d355349627aa0ab.

+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/dash/sources/coinmarketcap/CMCTicker.java
@@ -50,7 +50,7 @@ public class CMCTicker {
         this.symbol = symbol;
     }
 
-    public int getRank() {
+    public BigDecimal getRank() {
         return rank;
     }
